+++ pymc3/tests/test_distributions.py
@@ -309,8 +309,10 @@ def logpow(v, p):
 
 
 def discrete_weibull_logpmf(value, q, beta):
-    return floatX(np.log(np.power(q, np.power(value, beta))
-                  - np.power(q, np.power(value + 1, beta))))
+    return floatX(np.log(np.power(floatX(q),
+                                  np.power(floatX(value), floatX(beta)))
+                  - np.power(floatX(q), np.power(floatX(value + 1),
+                                                 floatX(beta)))))
 
 
 def dirichlet_logpdf(value, a):
@@ -565,11 +567,6 @@ class TestMatchesScipy(SeededTest):
         def scipy_logp(value, mu, sigma, lower, upper):
             return sp.truncnorm.logpdf(
                 value, (lower-mu)/sigma, (upper-mu)/sigma, loc=mu, scale=sigma)
-
-        args = {'mu': array(-2.1), 'lower': array(-100.), 'upper': array(0.01),
-                'sigma': array(0.01)}
-        val = TruncatedNormal.dist(**args).logp(0.)
-        assert_allclose(val.eval(), scipy_logp(value=0, **args))
 
         self.pymc3_matches_scipy(
             TruncatedNormal, R,
@@ -1292,16 +1289,16 @@ class TestLatex:
             sigma = HalfNormal('sigma', sigma=1)
 
             # Expected value of outcome
-            mu = Deterministic('mu', alpha + tt.dot(X, b))
+            mu = Deterministic('mu', floatX(alpha + tt.dot(X, b)))
 
             # Likelihood (sampling distribution) of observations
             Y_obs = Normal('Y_obs', mu=mu, sigma=sigma, observed=Y)
         self.distributions = [alpha, sigma, mu, b, Y_obs]
         self.expected = (
-            r'$\text{alpha} \sim \text{Normal}(\mathit{mu}=0,~\mathit{sigma}=10.0)$',
+            r'$\text{alpha} \sim \text{Normal}(\mathit{mu}=0.0,~\mathit{sigma}=10.0)$',
             r'$\text{sigma} \sim \text{HalfNormal}(\mathit{sigma}=1.0)$',
             r'$\text{mu} \sim \text{Deterministic}(\text{alpha},~\text{Constant},~\text{beta})$',
-            r'$\text{beta} \sim \text{Normal}(\mathit{mu}=0,~\mathit{sigma}=10.0)$',
+            r'$\text{beta} \sim \text{Normal}(\mathit{mu}=0.0,~\mathit{sigma}=10.0)$',
             r'$\text{Y_obs} \sim \text{Normal}(\mathit{mu}=\text{mu},~\mathit{sigma}=f(\text{sigma}))$'
         )